finalize functions (meta info) refactored
