+++ witness/gen.go
@@ -1,12 +1,10 @@
 package witness
 
 import (
-	"encoding/binary"
 	"fmt"
 	"log"
 	"math/big"
 	"os"
-	"strconv"
 
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/crypto"
@@ -91,7 +89,7 @@ const (
 	StorageLeafKeyCRow
 	AccountLeafKeyCRow // 4
 	HashRow
-	AccountLeafKeyRow // 6
+	AccountLeafKeySRow // 6
 	// TODO
 )
 
@@ -117,39 +115,6 @@ type TrieModification struct {
 	CodeHash []byte
 }
 
-func check(err error) {
-	if err != nil {
-		log.Fatal(err)
-	}
-}
-
-func MatrixToJson(rows [][]byte) string {
-	// Had some problems with json.Marshal, so I just prepare json manually.
-	json := "["
-	for i := 0; i < len(rows); i++ {
-		json += listToJson(rows[i])
-		if i != len(rows)-1 {
-			json += ","
-		}
-	}
-	json += "]"
-
-	return json
-}
-
-func listToJson(row []byte) string {
-	json := "["
-	for j := 0; j < len(row); j++ {
-		json += strconv.Itoa(int(row[j]))
-		if j != len(row)-1 {
-			json += ","
-		}
-	}
-	json += "]"
-
-	return json
-}
-
 // isBranch takes GetProof element and returns whether the element is a branch.
 func isBranch(proofEl []byte) bool {
 	elems, _, err := rlp.SplitList(proofEl)
@@ -160,58 +125,6 @@ func isBranch(proofEl []byte) bool {
 		log.Fatal("Proof element is neither leaf or branch")
 	}
 	return c == 17
-}
-
-// Equip proof with intermediate state roots, first level info, counter, address RLC,
-// modification tag (whether it is storage / nonce / balance change).
-func insertMetaInfo(stream, sRoot, cRoot, address, counter []byte, notFirstLevel, isStorageMod, isNonceMod, isBalanceMod, isCodeHashMod, isAccountDeleteMod, isNonExistingAccount, isNonExistingStorage byte) []byte {
-	// The last byte (-1) in a row determines the type of the row.
-	// Byte -2 determines whether it's the first level or not.
-	// Bytes before that store intermediate final and end roots.
-	l := len(stream)
-	extendLen := 64 + 32 + 32 + counterLen + 1 + 7
-	extended := make([]byte, l + extendLen) // make space for 32 + 32 + 32 + 1 (s hash, c hash, public_root, notFirstLevel)
-	copy(extended, stream)
-	extended[l+extendLen-1] = extended[l-1] // put selector to the last place
-	for i := 0; i < len(sRoot); i++ {
-		extended[l-1+i] = sRoot[i]
-	}
-	for i := 0; i < len(cRoot); i++ {
-		extended[l-1+len(sRoot)+i] = cRoot[i]
-	}
-	for i := 0; i < len(address); i++ {
-		extended[l-1+len(sRoot) + len(cRoot) + i] = address[i]
-	}
-	for i := 0; i < len(counter); i++ {
-		extended[l-1+len(sRoot) + len(cRoot) + len(address) + i] = counter[i]
-	}
-	// public root set later
-
-	extended[l+extendLen-2] = notFirstLevel
-	extended[l+extendLen-3] = isStorageMod
-	extended[l+extendLen-4] = isNonceMod
-	extended[l+extendLen-5] = isBalanceMod
-	extended[l+extendLen-6] = isCodeHashMod
-	extended[l+extendLen-7] = isAccountDeleteMod
-	extended[l+extendLen-8] = isNonExistingAccount
-	extended[l+extendLen-9] = isNonExistingStorage
-
-	return extended
-}
-
-func insertPublicRoot(proof [][]byte, startRoot, finalRoot []byte) {
-	for i := 0; i < len(proof); i++ {
-		l := len(proof[i])
-		if i == 0 {
-			for j := 0; j < 32; j++ {
-				proof[i][l - 32 - 6 + j] = startRoot[j]
-			}
-		} else {
-			for j := 0; j < 32; j++ {
-				proof[i][l - 32 - 6 + j] = finalRoot[j]
-			}
-		}
-	}
 }
 
 // addForHashing takes the stream of bytes and append a byte to it that marks this stream
@@ -241,58 +154,6 @@ func GetParallelProofs(nodeUrl string, blockNum int, trieModifications []TrieMod
 	}
 
 	return getParallelProofs(trieModifications, statedb, 0)
-}
-
-func prepareProof(ind int, newProof [][]byte, addrh []byte, sRoot, cRoot, startRoot, finalRoot common.Hash, mType ModType) [][]byte {
-	firstLevelBoundary := branchRows
-	if newProof[0][len(newProof[0])-1] == 6 {
-		// 6 presents account leaf key S.
-		// This happens when account leaf is without branch / extension node.
-		firstLevelBoundary = accountLeafRows
-	}
-
-	isStorageMod := byte(0)
-	isNonceMod := byte(0)
-	isBalanceMod := byte(0)
-	isCodeHashMod := byte(0)
-	isAccountDeleteMod := byte(0)
-	isNonExistingAccount := byte(0)
-	isNonExistingStorage := byte(0)
-	if mType == StorageMod {
-		isStorageMod = 1
-	} else if mType == NonceMod {
-		isNonceMod = 1
-	} else if mType == BalanceMod {
-		isBalanceMod = 1
-	} else if mType == CodeHashMod {
-		isCodeHashMod = 1
-	} else if mType == CreateAccount {
-		isNonceMod = 1 // TODO: setting as nonce mod for now, this depends on the lookup
-	} else if mType == DeleteAccount {
-		isAccountDeleteMod = 1
-	} else if mType == NonExistingAccount {
-		isNonExistingAccount = 1
-	} else if mType == NonExistingStorage {
-		isNonExistingStorage = 1
-	}
-
-	counter := make([]byte, counterLen)
-	binary.BigEndian.PutUint32(counter[0:4], uint32(ind))
-
-	proof := [][]byte{}
-	for j := 0; j < len(newProof); j++ {
-		notFirstLevel := byte(1)
-		if j < firstLevelBoundary {
-			notFirstLevel = 0
-		}
-		r := insertMetaInfo(newProof[j], sRoot.Bytes(), cRoot.Bytes(), addrh, counter, notFirstLevel, 
-			isStorageMod, isNonceMod, isBalanceMod, isCodeHashMod, isAccountDeleteMod, isNonExistingAccount,
-			isNonExistingStorage)
-		proof = append(proof, r)
-	}
-	insertPublicRoot(newProof, startRoot.Bytes(), finalRoot.Bytes())
-
-	return proof
 }
 
 func prepareAccountProof(i int, tMod TrieModification, tModsLen int, statedb *state.StateDB, specialTest byte) ([][]byte, [][]byte) {
@@ -369,7 +230,7 @@ func prepareAccountProof(i int, tMod TrieModification, tModsLen int, statedb *st
 	rowsState, toBeHashedAcc, _ :=
 		prepareWitness(statedb, addr, accountProof, accountProof1, aExtNibbles1, aExtNibbles2, accountAddr, aNode, true, tMod.Type == NonExistingAccount, false, isShorterProofLastLeaf)
 
-	proof := prepareProof(i, rowsState, addrh, sRoot, cRoot, startRoot, finalRoot, tMod.Type)
+	proof := finalizeProof(i, rowsState, addrh, sRoot, cRoot, startRoot, finalRoot, tMod.Type)
 
 	return proof, toBeHashedAcc
 }
@@ -456,7 +317,7 @@ func getParallelProofs(trieModifications []TrieModification, statedb *state.Stat
 				prepareWitness(statedb, addr, storageProof, storageProof1, extNibbles1, extNibbles2, keyHashed, node, false, false, tMod.Type == NonExistingStorage, isLastLeaf)
 			rowsState = append(rowsState, rowsStorage...)
 	
-			proof := prepareProof(i, rowsState, addrh, sRoot, cRoot, startRoot, finalRoot, tMod.Type)
+			proof := finalizeProof(i, rowsState, addrh, sRoot, cRoot, startRoot, finalRoot, tMod.Type)
 			allProofs = append(allProofs, proof...)
 			
 			// Put rows that just need to be hashed at the end, because circuit assign function
@@ -531,4 +392,161 @@ func UpdateStateAndGenProof(testName string, keys, values []common.Hash, address
 	n3, err := f.WriteString(w)
     check(err)
     fmt.Printf("wrote %d bytes\n", n3)
+}
+
+// prepareWitness takes two GetProof proofs (before and after single modification) and prepares
+// a witness for the MPT circuit. Alongside, it prepares the byte streams that need to be hashed
+// and inserted into Keccak lookup table.
+func prepareWitness(statedb *state.StateDB, addr common.Address, proof1, proof2, extNibblesS, extNibblesC [][]byte, key []byte, neighbourNode []byte,
+		isAccountProof, nonExistingAccountProof, nonExistingStorageProof, isShorterProofLastLeaf bool) ([][]byte, [][]byte, bool) {
+	rows := make([][]byte, 0)
+	toBeHashed := make([][]byte, 0)
+
+	minLen := len(proof1)
+	if len(proof2) < minLen {
+		minLen = len(proof2)
+	}
+
+	keyIndex := 0
+	len1 := len(proof1)
+	len2 := len(proof2)
+
+	// When a value in the trie is updated, both proofs are of the same length.
+	// Otherwise, when a value is added (not updated) and there is no node which needs to be changed
+	// into a branch, one proof has a leaf and one does not have it.
+	// The third option is when a value is added and the existing leaf is turned into a branch,
+	// in this case we have an additional branch in C proof (when deleting a value causes
+	// that a branch with two leaves turns into a leaf, we have an additional branch in S proof).
+
+	additionalBranch := false
+	if len1 < len2 && len1 > 0 { // len = 0 when trie trie is empty
+		// Check if the last proof element in the shorter proof is a leaf -
+		// if it is, then there is an additional branch.
+		additionalBranch = !isBranch(proof1[len1 - 1])
+	} else if len2 < len1 && len2 > 0 {
+		additionalBranch = !isBranch(proof2[len2 - 1])
+	}
+
+	upTo := minLen
+	if (len1 != len2) && additionalBranch {
+		upTo = minLen - 1
+	}
+
+	var extensionRowS []byte
+	var extensionRowC []byte
+	extensionNodeInd := 0
+
+	branchC16 := byte(0); 
+	branchC1 := byte(1);
+	for i := 0; i < upTo; i++ {
+		if !isBranch(proof1[i]) {
+			if i != len1 - 1 { // extension node
+				var numberOfNibbles byte
+				numberOfNibbles, extensionRowS, extensionRowC = prepareExtensionRows(extNibblesS, extensionNodeInd, proof1[i], proof2[i], false, false)
+				keyIndex += int(numberOfNibbles)
+				extensionNodeInd++
+				continue
+			}
+
+			l := len(proof1)
+			addLeafRows(&rows, proof1[l-1], proof2[l-1], key, nonExistingAccountProof, nonExistingStorageProof, isAccountProof, &toBeHashed)
+		} else {
+			switchC16 := true // If not extension node, switchC16 = true.
+			if extensionRowS != nil {
+				keyLen := getExtensionNodeKeyLen(proof1[i-1])
+				if keyLen == 1 {
+					switchC16 = false
+				} else {
+					if proof1[i-1][2] != 0 { // If even, switch16 = true.
+						switchC16 = false
+					}
+				}
+			}
+			if switchC16 {
+				if branchC16 == 1 {
+					branchC16 = 0
+					branchC1 = 1
+				} else {
+					branchC16 = 1
+					branchC1 = 0
+				}
+			}
+
+			bRows := prepareTwoBranches(proof1[i], proof2[i], key[keyIndex], branchC16, branchC1, false, false)
+			keyIndex += 1
+
+			// extension node rows
+			if extensionRowS != nil {
+				bRows = append(bRows, extensionRowS)
+				bRows = append(bRows, extensionRowC)
+
+				setExtensionNodeSelectors(&bRows, proof1[i-1], proof2[i-1], branchC16, branchC1)
+
+				// adding extension nodes for hashing:
+				addForHashing(proof1[i-1], &toBeHashed)
+				addForHashing(proof2[i-1], &toBeHashed)
+			} else {
+				extRows := prepareEmptyExtensionRows(false, false)
+				bRows = append(bRows, extRows...)
+			}
+
+			rows = append(rows, bRows...)
+			addForHashing(proof1[i], &toBeHashed)
+			addForHashing(proof2[i], &toBeHashed)
+
+			extensionRowS = nil
+			extensionRowC = nil
+		}
+	}	
+	
+	if len1 != len2 {
+		if additionalBranch {	
+			var leafRows [][]byte
+			var leafForHashing [][]byte
+			if isAccountProof {
+				leafRows, leafForHashing = prepareAccountLeaf(proof1[len1-1], proof2[len2-1], key, nonExistingAccountProof, false)
+				toBeHashed = append(toBeHashed, leafForHashing...)
+			} else {
+				leafRows, leafForHashing = prepareStorageLeaf(proof1[len1-1], key, nonExistingAccountProof)
+				toBeHashed = append(toBeHashed, leafForHashing...)
+			}
+			
+			isModifiedExtNode, isExtension, numberOfNibbles, branchC16 := addBranchAndPlaceholder(statedb, addr, &rows, proof1, proof2, extNibblesS, extNibblesC,
+				leafRows[0], key, neighbourNode,
+				keyIndex, extensionNodeInd, additionalBranch,
+				isAccountProof, nonExistingAccountProof, isShorterProofLastLeaf, branchC16, branchC1, &toBeHashed)
+
+			if isAccountProof {
+				addAccountLeafAfterBranchPlaceholder(&rows, proof1, proof2, leafRows, neighbourNode, key, nonExistingAccountProof, isModifiedExtNode, isExtension, numberOfNibbles, &toBeHashed)	
+			} else {	
+				addStorageLeafAfterBranchPlaceholder(&rows, proof1, proof2, leafRows, neighbourNode, key, nonExistingAccountProof, isModifiedExtNode, isExtension, numberOfNibbles, &toBeHashed)
+			}
+
+			// When a proof element is a modified extension node (new extension node appears at the position
+			// of the existing extension node), additional rows are added (extension node before and after
+			// modification).
+			if isModifiedExtNode {
+				addModifiedExtNode(statedb, addr, &rows, proof1, proof2, extNibblesS, extNibblesC, key, neighbourNode,
+					keyIndex, extensionNodeInd, numberOfNibbles, additionalBranch,
+					isAccountProof, nonExistingAccountProof, isShorterProofLastLeaf, branchC16, branchC1, &toBeHashed)
+			}
+		} else {
+			addLeafAndPlaceholder(&rows, proof1, proof2, key, nonExistingAccountProof, isAccountProof, &toBeHashed)
+		}
+	} else if isBranch(proof2[len(proof2)-1]) {
+		// Let's always use C proof for non-existing proof.
+		// Account proof has drifted leaf as the last row, storage proof has non-existing-storage row
+		// as the last row.
+		// When non existing proof and only the branches are returned, we add a placeholder leaf.
+		// This is to enable the lookup (in account leaf row), most constraints are disabled for these rows.
+		if isAccountProof {
+			leafRows := prepareAccountLeafPlaceholderRows(key, keyIndex, nonExistingAccountProof)
+			rows = append(rows, leafRows...)	
+		} else {
+			leafRows := prepareStorageLeafPlaceholderRows(key, keyIndex, nonExistingStorageProof)
+			rows = append(rows, leafRows...)	
+		}
+	}
+
+	return rows, toBeHashed, extensionNodeInd > 0
 }
